bugfix.ts refactor add applyFix.test.ts with unit test
